rename module to Data.Text.I18n.
- update documentation and specs.
- remove redundant test module.
- rename i18n modules.
- update cabal file.
- update makefile with new path.
- extract common types module.
- reformat imports.
- add documentation for Data.Text.I18n.Types.
