+++ i18n.cabal
@@ -14,8 +14,9 @@ cabal-version:       >=1.10
 
 library
     hs-source-dirs:    src
-    exposed-modules:   Text.I18n
-                     , Text.I18n.Po
+    exposed-modules:   Data.Text.I18n
+                     , Data.Text.I18n.Po
+                     , Data.Text.I18n.Types
     build-depends:     base >= 4.7 && < 5
                      , containers
                      , filepath
